refactor: split ComponentManager types mechanically
