+++ Editor/ComponentManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5567da64b912355352c9a3a6b605ad50
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
