+++ Editor/ComponentManager/ComponentConstants.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 820b79ba28c53ad516eefe0ca0059efd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
