+++ Editor/ComponentManager/ComponentDataManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6da4118e7e96f6be54a5651eab1f399c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
