+++ Editor/ComponentManager/ComponentInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9db72844598e695c3d4259f044a352b4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
