+++ Editor/ComponentManager/ComponentPathUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f4a05776a8318b8f70f2bd2b2d59d7c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
